+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@
 - change: rename TestCaseInOut field, config_vars and export_vars
 - change: rename StepData field, export_vars
 - change: add `--tb=short` for `hrun` command to use shorter traceback format by default
+- change: search debugtalk.py upward recursively until system root dir
 
 ## 3.0.8 (2020-06-04)